Fix test file names
